+++ app/src/main/java/com/jamie/hn/stories/ui/StoryViewItem.kt
@@ -1,7 +1,7 @@
 package com.jamie.hn.stories.ui
 
 data class StoryViewItem(
-    val id: Long,
+    val id: Int,
     val author: String,
     val comments: String,
     val score: String,
@@ -9,6 +9,6 @@ data class StoryViewItem(
     val time: String,
     val title: String,
     val url: String,
-    val commentsCallback: (Long) -> Unit,
-    val storyViewerCallback: (Long) -> Unit
+    val commentsCallback: (Int) -> Unit,
+    val storyViewerCallback: (Int) -> Unit
 )
